+++ ProjectProb/app/src/main/java/com/sharktech/projectprob/models/VariableString.java
@@ -95,8 +95,8 @@ public class VariableString implements IVariable {
         }
 
         @Override
-        public Long asNumber() {
-            return 1L;
+        public Float asNumber() {
+            return 1f;
         }
 
         @Override
